Re-organized kyrin's internal structure. Initiated integrated logger implementation using winston project.
