+++ package.json
@@ -7,7 +7,7 @@
     "start": "pm2 start env/kyrin-ecosystem.config.json",
     "build": "tsc",
     "stop": "pm2 kill",
-    "dev:verbose":"DEBUG=express* node boot/www.js"
+    "dev:verbose": "DEBUG=express* node boot/www.js"
   },
   "author": "Ravi Semwal",
   "license": "GPL-2.0",
@@ -31,17 +31,20 @@
     "typescript": "^2.1.5"
   },
   "dependencies": {
+    "@types/winston": "^2.2.0",
     "body-parser": "^1.16.0",
     "core-js": "^2.4.1",
     "ejs": "^2.5.5",
     "express": "^4.14.0",
     "express-session": "^1.15.0",
+    "express-winston": "^2.2.0",
     "helmet": "^3.4.0",
     "js-yaml": "^3.7.0",
     "mongoose": "^4.7.7",
     "passport": "^0.3.2",
     "pm2": "^2.4.0",
     "serve-favicon": "^2.3.2",
-    "socket.io": "^1.7.2"
+    "socket.io": "^1.7.2",
+    "winston": "^2.3.1"
   }
 }
